API/requirements.txt
tensorflow
keras
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy